move fetchAdvice logic to AdviceCard- better Divider
